pyproject.toml
[project]
name = "theloop"
version = "0.1.0"
description = "Upload files from URLs to cloud storage"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Vladimir Vlasiuk", email = "matterai.net@gmail.com" }]
keywords = ["cloud", "storage", "upload", "gcp", "cli"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "typer==0.15.4",
    "click>=8.1,<8.2",
    "tqdm==4.67.1",
    "requests==2.32.3",
    "google-cloud-storage==3.1.0",
    "aiohttp==3.11.18",
    "pydantic==2.11.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
theloop = "theloop.cli:app"

[project.urls]
Homepage = "https://github.com/matterai/theloop"
Repository = "https://github.com/matterai/theloop"
Issues = "https://github.com/matterai/theloop/issues"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-asyncio>=0.26.0",
]
